code/requirements.txt
PyGithub
pymongo==3.5.1